requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
bottleneck>=1.3.6

# Data handling
python-dotenv>=1.0.0
sqlalchemy>=2.0.0

# SEC data
sec-edgar-downloader>=5.0.0

# Market data
yfinance>=0.2.28

# Analysis
scikit-learn>=1.3.0

# Utilities
tqdm>=4.66.0
loguru>=0.7.0

# Visualization & Reporting
plotly>=5.14.0
rich>=13.5.0
streamlit>=1.28.0

# Optional: For notebooks
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
